Refactor main() into site specific and common code.
The actual main() entry point is reduced to some basic setup code that
can be substituted out as a module when more complex executable setup
is required with certain site's deployments.  Left in main() is
SignalAction initialization, Options parsing, and HotRestarter
allocation.  Everything else is moved into main_common().

A future commit will further shift the hot restart related code into
the site-specific main() so deployments not using hot restart do not
have to depend on the code.

This is a purely mechanical refactoring, no logic is changed.
